Replace the useEffect() with useLayoutEffect() #438
To ensure that the DOM element of the active DrawerPanel is fully rendered, it is recommended to use the useLayoutEffect() hook instead of useEffect(). The useLayoutEffect() hook runs after the DOM has been rendered, but before the browser has painted the screen, which makes it suitable for accessing DOM measurements such as element heights.